created generated address for 2 public screen
