+++ Frontend/lib/screens/home/home.dart
@@ -1,4 +1,4 @@
-export 'generate_address.dart';
+export 'generate_address_two.dart';
 export 'home_page.dart';
 export 'home_screen.dart';
 export 'set_up.dart';
